requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bibtexparser==1.2.0
Bootstrap-Flask==1.8.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
distlib==0.3.3
Django==3.2.9
djangorestframework==3.12.4
feedparser==6.0.8
filelock==3.2.0
Flask==2.0.2
Flask-Babel==2.0.0
Flask-WTF==1.0.0
future==0.18.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Kerko==0.8.1
MarkupSafe==2.0.1
Pillow==8.4.0
platformdirs==2.4.0
promise==2.3
pyparsing==3.0.6
pytz==2021.3
Pyzotero==1.4.26
requests==2.26.0
Rx==1.6.1
sgmllib3k==1.0.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
virtualenv==20.8.1
w3lib==1.22.0
Werkzeug==2.0.2
Whoosh==2.7.4
wrapt==1.13.3
WTForms==2.3.3
